Extract UpDownBase nested classes
