+++ src/System.Windows.Forms/src/System/Windows/Forms/UpDownBase.ButtonID.cs
@@ -1,0 +1,16 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.Windows.Forms
+{
+    public abstract partial class UpDownBase
+    {
+        internal enum ButtonID
+        {
+            None = 0,
+            Up = 1,
+            Down = 2,
+        }
+    }
+}
